final fix

+++ src/backend/models/Recipe.php
@@ -259,7 +259,9 @@
       //bind data
       $stmt->bindParam(':recipe_id', $this->recipe_id);
       $stmt->execute();
-
+      if($stmt->rowCount()<1) {
+        return false;
+      }
       //delete from ingredients
       $query = 'DELETE FROM ingredient
                 WHERE recipe_id = :recipe_id';
@@ -412,7 +414,11 @@
       $stmt->execute();
       $row = $stmt->fetch();
       // echo $row['user_id'];
-      return $row['user_id'];
+      if($stmt->rowCount()>=1) {
+        return $row['user_id'];
+      } else {
+        return false;
+      }
     }
 
     public function checkCommentOwnership($id){
@@ -539,11 +545,30 @@
       if($this->conn->commit()){
         return true;
       }
-
-      //print error if something goes wrong
-      // printf('Error: %s.\n', $stmt->error);
-
       return false;
-    
+    }
+
+    public function logging($action, $success) {
+      $query = 'INSERT INTO log
+                SET
+                ip = :ip,
+                browser = :browser,
+                action = :action,
+                success= :success';
+
+      //prepate statement
+      $stmt = $this->conn->prepare($query);
+
+      //bind data
+      $stmt->bindParam(':ip', $_SERVER['REMOTE_ADDR']);
+      $stmt->bindParam(':browser', $_SERVER['HTTP_USER_AGENT']);
+      $stmt->bindParam(':action', $action);
+      $stmt->bindParam(':success', $success);
+
+      //save into database
+      if($stmt->execute()){
+        return true;
+      }
+      return false;
     }
   }
